Split modal code
